+++ app/controllers/intouch_controller.rb
@@ -1,7 +1,7 @@
 class IntouchController < ApplicationController
   unloadable
 
-  before_filter :find_project
+  before_filter :find_project, only: [:save_settings]
 
   def save_settings
     if request.put?
@@ -13,6 +13,19 @@ class IntouchController < ApplicationController
     end
 
     redirect_to controller: 'projects', action: 'settings', tab: params[:tab] || 'intouch_settings', id: @project
+  end
+
+  def bot_init
+    bot = Telegram::Bot::Client.new(Intouch.bot_token)
+    bot.api.setWebhook(url: Intouch.web_hook_url)
+
+    redirect_to plugin_settings_path('redmine_intouch'), notice: t('intouch.bot.authorize.success')
+  end
+
+  def bot_deinit
+    bot = Telegram::Bot::Client.new(Intouch.bot_token)
+    bot.api.setWebhook(url: '')
+    redirect_to plugin_settings_path('redmine_intouch'), notice: t('intouch.bot.deauthorize.success')
   end
 
   private