+++ lib/intouch.rb
@@ -10,6 +10,10 @@ module Intouch
     Setting.plugin_redmine_intouch['telegram_bot_token']
   end
 
+  def self.web_hook_url
+    "https://#{Setting.host_name}/intouch/web_hook"
+  end
+
   def self.sidekiq_cron_jobs
     names = %w(cron_overdue_regular_notification cron_working_regular_notification cron_unassigned_regular_notification cron_feedback_regular_notification)
     Sidekiq::Cron::Job.all.select { |job| names.include? job.name }
@@ -21,6 +25,10 @@ module Intouch
 
   def self.active_protocols
     Setting.plugin_redmine_intouch['active_protocols'] || []
+  end
+
+  def self.handle_message(message)
+    Intouch::TelegramBot.new(message).call if message.is_a?(Telegram::Bot::Types::Message)
   end
 
   def self.available_recipients